feat: add RuntimeUsageEntryCard component and runtime usage dialog model
- Implemented RuntimeUsageEntryCard component for displaying runtime usage details.
- Created runtimeUsageDialogModel.ts to define RuntimeUsageEntry interface and related utility functions.
- Added formatting and data handling functions for runtime limits, usage summaries, and provider metadata.
- Integrated model and tool usage summaries with detailed breakdowns in the UI.
