+++ packages/shared/src/browser.ts
@@ -67,6 +67,7 @@ export {
   buildRuntimeLimitSignature,
   normalizeRuntimeLimitSnapshot,
   redactProviderText,
+  redactProviderTextForLogs,
   resolveRuntimeLimitFutureHint,
   sanitizeRuntimeLimitSnapshotForExposure,
   sanitizeProviderMeta,
